Project51 (Pentester Quick Tools Guide)
So long story short: This tool will act as a quick guide for Pentester (beginner to intermediate), the project consists of 51 tools guide set with proper usage, description, and syntax (some having walkthrough too). So, for example, if Pentester wants to perform Android Hacking/Testing, they will check the  Mobile Hacking guide and be provided with a bunch of tools helpful to perform testing or checks with usage. This way, they can save time struggling to find available tools for specific work and searching for manuals of cheat sheets for those standard tools.